requirements.txt
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pandas==2.3.1
playwright==1.53.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pyjokes==0.8.3
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-timeout==2.4.0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
SQLAlchemy==2.0.41
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1